+++ tests/test_aeat_functional.py
@@ -1,8 +1,7 @@
 import pytest
 
-from tests import factories
-
 import aeat
+from aeat.utils import factories
 
 
 '''
